GC: introduced the ability to retire TLAB in the HeapSpace interface and extended the Evacuator interface to enable setting a evacuation buffer size. Used these for specific aspects of GenSSHeapScheme implementation's minor collection: minor collection allocate an evacuation buffer sized to the free space left in the old generation's to space, and retire the left over at the end of minor collection.

+++ com.oracle.max.vm/src/com/sun/max/vm/heap/gcx/mse/MSEHeapScheme.java
@@ -195,7 +195,7 @@ public final class MSEHeapScheme extends HeapSchemeWithTLABAdaptor implements He
     public boolean collectGarbage(Size requestedFreeSpace) {
         final Size usedSpaceBefore = markSweepSpace.usedSpace();
         if (MaxineVM.isDebug()) {
-            final int logCursor = (int) (collectionCount % 16);
+            final int logCursor = collectionCount % 16;
             allocatedSinceLastGC[logCursor] = usedSpaceBefore.minus(usedSpaceAfterLastGC).toLong();
             if (logCursor == 15) {
                 long c = collectionCount - 15;
